@@ -1,47 +1,37 @@
 {
   "name": "smee-client",
-  "version": "1.2.2",
+  "version": "2.0.0",
   "description": "Client to proxy webhooks to local host",
-  "main": "index.js",
+  "main": "./bin/index.js",
+  "private": true,
   "bin": {
-    "smee": "./bin/smee.js"
+    "smee": "./bin/index.js"
   },
   "files": [
-    "index.js",
-    "index.d.ts",
-    "bin",
-    "lib"
+    "bin"
   ],
   "scripts": {
     "test": "jest --coverage && standard",
     "build": "tsc -p tsconfig.json"
   },
-  "repository": "github:probot/smee-client",
+  "repository": "github:ethomson/smee-client",
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "commander": "^2.19.0",
-    "eventsource": "^1.0.7",
-    "morgan": "^1.9.1",
-    "superagent": "^5.0.2",
+    "commander": "^9.0.0",
+    "eventsource": "^2.0.0",
+    "morgan": "^1.10.0",
+    "superagent": "^7.1.1",
     "validator": "^13.7.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.4.0",
-    "@types/eventsource": "^1.1.0",
-    "@types/jest": "^24.0.11",
-    "@types/nock": "^9.3.1",
-    "@types/superagent": "^4.1.1",
-    "@types/validator": "^10.11.0",
-    "babel-core": "^7.0.0-bridge.0",
-    "babel-jest": "^24.5.0",
-    "connect-sse": "^1.2.0",
-    "jest": "^24.5.0",
-    "nock": "^10.0.6",
-    "standard": "^12.0.1",
-    "supertest": "^4.0.2",
-    "ts-jest": "^24.0.1",
-    "typescript": "^3.4.1"
+    "@types/eventsource": "^1.1.8",
+    "@types/jest": "^27.4.1",
+    "@types/superagent": "^4.1.15",
+    "@types/validator": "^13.7.1",
+    "nock": "^13.2.4",
+    "ts-jest": "^27.1.3",
+    "typescript": "^4.6.2"
   },
   "standard": {
     "env": [
